pyproject.toml
[project]
name = "geekmagic-hacs"
version = "1.0.0"
description = "Home Assistant custom integration for GeekMagic displays"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
authors = [
    { name = "Adrien Brault" }
]
dependencies = [
    "pillow>=10.0.0",
    "aiohttp>=3.8.0",
    "palettable>=3.3.0",
    "stretchable>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-aiohttp>=1.0.0",
    "pytest-cov>=4.0.0",
    "homeassistant>=2024.10.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["custom_components/geekmagic"]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-aiohttp>=1.0.0",
    "pytest-cov>=4.0.0",
    "pytest-homeassistant-custom-component>=0.13.0",
    "ruff>=0.8.0",
    "homeassistant>=2024.10.0",
    "voluptuous>=0.13.0",
    "ty>=0.0.1a7",
    "pre-commit>=4.0.0",
    # Experimental rendering libraries
    "aggdraw>=1.3.0",
    "pycairo>=1.26.0",
    "palettable>=3.3.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "A",      # flake8-builtins
    "ASYNC",  # flake8-async
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "DTZ",    # flake8-datetimez
    "E",      # pycodestyle errors
    "F",      # pyflakes
    "FA",     # flake8-future-annotations
    "FLY",    # flynt
    "FURB",   # refurb
    "I",      # isort
    "ICN",    # flake8-import-conventions
    "INT",    # flake8-gettext
    "LOG",    # flake8-logging
    "N",      # pep8-naming
    "PERF",   # perflint
    "PGH",    # pygrep-hooks
    "PIE",    # flake8-pie
    "PL",     # pylint
    "PT",     # flake8-pytest-style
    "PTH",    # flake8-use-pathlib
    "PYI",    # flake8-pyi
    "Q",      # flake8-quotes
    "RET",    # flake8-return
    "RSE",    # flake8-raise
    "RUF",    # ruff-specific
    "S",      # flake8-bandit (security)
    "SIM",    # flake8-simplify
    "SLF",    # flake8-self
    "SLOT",   # flake8-slots
    "T10",    # flake8-debugger
    "TCH",    # flake8-type-checking
    "TID",    # flake8-tidy-imports
    "TRY",    # tryceratops
    "UP",     # pyupgrade
    "W",      # pycodestyle warnings
    "YTT",    # flake8-2020
]
ignore = [
    "PLR0912",  # too-many-branches (icon drawing dispatch)
    "PLR0913",  # too-many-arguments
    "PLR0915",  # too-many-statements (icon drawing, sample generators)
    "PLR2004",  # magic-value-comparison
    "S101",     # assert (used in tests)
    "TC001",    # type-checking imports (HA uses at runtime)
    "TC002",    # type-checking imports (HA uses at runtime)
    "TID252",   # relative imports from parent (common in HA integrations)
    "TRY003",   # raise-vanilla-args
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "PLC0415",  # imports in fixtures
    "PT018",    # compound assertions OK in tests
    "S101",     # assert allowed in tests
    "SLF001",   # private member access in tests
    "PLR2004",  # magic values in tests
]
"custom_components/geekmagic/widgets/component_helpers.py" = [
    "N802",     # PascalCase for component factory functions (React-style)
]

[tool.ruff.lint.isort]
known-first-party = ["custom_components.geekmagic"]

[tool.coverage.run]
source = ["custom_components/geekmagic"]

[tool.ty.environment]
python-version = "3.12"
